Runtime: no require with target-env browser
