+++ toplevel/examples/lwt_toplevel/dune
@@ -90,6 +90,8 @@
    compile
    --pretty
    --Werror
+   --target-env
+   browser
    -I
    .
    --file
